+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - (New) Objective class: a class that represents an objective function of a problem. Each Objective has a name, a direction (minimize or maximize), and functions for extracting the value of the objective from Moves and from Solutions. See docs for a detailed description.
 - (New) Util methods: CollectionUtil::getBest(Iterable,ToDoubleFunction), ArrayUtil::stats(double[]), ArrayUtil::stats(int[]), ArrayUtil::stats(long[]), ReferenceResult::setScore(String, double)
 - (New) Validator methods: getFailCount, getReasonsFailed
+- (New) Allow users to enable or disable validations for the current thread. New methods: Context.Configurator.enableValidations(), Context.Configurator.disableValidations()
 - (New) Support for AspectJ, first magic annotation @TimeStats implemented to measure the time of any method.
 - (New) Ablation test: run ablation test by default after autoconfig procedure.
 - (New) Irace plots: generate an HTML report with the results of running the autoconfig procedure.